Implements v0.10.6 API for Python
