--- src/renderer/features/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "python")]
-pub mod python;
