+++ src/renderer/platform/python/mod.rs
@@ -1,114 +1,141 @@
-use crate::{FragmentColorError, Frame, Pass, Renderer, Shader, Target};
+use crate::{FragmentColorError, Frame, Pass, Renderer, Shader};
+use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
-use pyo3::types::PyDict;
+use pyo3::types::{PyDict, PyTuple};
 
 pub mod target;
 pub use target::*;
 
-#[pyclass(dict)]
-#[derive(Debug)]
-pub struct RenderCanvasContext {
-    pub(crate) canvas: PyObject,
-    _present_methods: PyObject, // @TODO figure how RenderCanvas expects me to use this
+pub mod iterator;
+pub use iterator::*;
 
-    // Reference-counted for Python.
-    renderer: Option<Py<Renderer>>,
-    target: Option<Py<RenderCanvasTarget>>,
-}
+pub mod renderable;
+pub use renderable::*;
 
-fn clone_py<T>(object: &Py<T>) -> Py<T> {
-    Python::with_gil(|py| object.clone_ref(py))
-}
+pub mod handle;
+use handle::*;
 
 #[pymethods]
-impl RenderCanvasContext {
+impl Renderer {
     #[new]
-    pub fn new(canvas: PyObject, _present_methods: PyObject) -> Self {
-        RenderCanvasContext {
-            canvas,
-            _present_methods,
-            renderer: None,
-            target: None,
-        }
+    /// Creates an uninitialized Renderer.
+    ///
+    /// At this point we don't know if it should be compatible
+    /// with a headless or a windowed environment.
+    ///
+    /// The render context is initialized when we crate the first target.
+    pub fn new_py() -> Renderer {
+        Self::new()
     }
 
-    pub fn is_ready(&self) -> bool {
-        self.renderer.is_some() && self.target.is_some()
-    }
+    #[pyo3(name = "render")]
+    /// Renders a Frame or Shader to a Target.
+    pub fn render_py(
+        &self,
+        renderable: PyObject,
+        target: &RenderCanvasTarget,
+    ) -> Result<(), PyErr> {
+        Python::with_gil(|py| -> Result<(), PyErr> {
+            let renderable_type = renderable
+                .call_method0(py, "renderable_type")?
+                .extract::<String>(py)?;
 
-    // @TODO consider storing them as PyObject to be able to return them directly
-    #[getter]
-    pub fn renderer(&self) -> PyResult<Py<Renderer>> {
-        if let Some(renderer) = &self.renderer {
-            Ok(clone_py::<Renderer>(renderer))
-        } else {
-            Err(FragmentColorError::new_err("Renderer not initialized"))
-        }
-    }
+            match renderable_type.as_str() {
+                "Frame" => {
+                    let frame = renderable.bind(py).downcast::<Frame>()?;
+                    let renderable = PyRenderable::from(frame);
 
-    #[getter]
-    pub fn target(&self) -> PyResult<Py<RenderCanvasTarget>> {
-        if let Some(target) = &self.target {
-            Ok(clone_py::<RenderCanvasTarget>(target))
-        } else {
-            Err(FragmentColorError::new_err("Target not initialized"))
-        }
-    }
+                    self.render(&renderable, target)?;
 
-    // duck-typed interface that a context must implement, to be usable with RenderCanvas.
-    // Upstream documentation: https://rendercanvas.readthedocs.io/stable/contextapi.html
-    //
-    // fn canvas(&self) -> PyObject;
-    // fn present(&self) -> Result<Py<PyDict>, PyErr>;
-    //
-    // We can't export a impl Trait block with Pyo3.
-
-    #[getter]
-    pub fn canvas(&self) -> PyObject {
-        clone_py::<PyAny>(&self.canvas)
-    }
-
-    pub fn present(&self) -> Result<Py<PyDict>, PyErr> {
-        Python::with_gil(|py| -> PyResult<Py<PyDict>> {
-            let dict = PyDict::new(py);
-
-            if let Some(py_target) = &self.target {
-                let target = py_target.borrow(py);
-                match target.get_current_frame() {
-                    Ok(frame) => {
-                        frame.present();
-                        dict.set_item("method", "screen")?;
-                    }
-                    Err(e) => {
-                        dict.set_item("method", "fail")?;
-                        dict.set_item("message", e.to_string())?;
-                    }
+                    Ok(())
                 }
-            } else {
-                dict.set_item("method", "fail")?;
-                dict.set_item("message", "Target not initialized")?;
+                "Pass" => {
+                    let pass = renderable.bind(py).downcast::<Pass>()?;
+                    let renderable = PyRenderable::from(pass);
+
+                    self.render(&renderable, target)?;
+
+                    Ok(())
+                }
+                "Shader" => {
+                    let shader = renderable.bind(py).downcast::<Shader>()?;
+                    let renderable = PyRenderable::from(shader);
+
+                    self.render(&renderable, target)?;
+
+                    Ok(())
+                }
+                _ => Err(PyErr::new::<PyTypeError, _>(
+                    "Expected a Frame, Pass or Shader object",
+                )),
+            }
+        })?;
+
+        Ok(())
+    }
+
+    pub fn create_target(&self, rendercanvas: PyObject) -> Result<Py<RenderCanvasTarget>, PyErr> {
+        Python::with_gil(|py| -> Result<Py<RenderCanvasTarget>, PyErr> {
+            // If the target is already initialized, return it
+            let libname = PyTuple::new(py, &["fragmentcolor"])?;
+            let py_target = rendercanvas.call_method1(py, "get_context", libname)?; // calls hook
+            let bound_target = py_target.downcast_bound::<RenderCanvasTarget>(py)?;
+            let mut target = bound_target.borrow_mut();
+            if target.is_ready() {
+                return Ok(target.into_pyobject(py)?.unbind());
+            }
+
+            // Returns a list of the possible present methods ("screen", "bitmap")
+            let present_methods = rendercanvas.call_method0(py, "_rc_get_present_methods")?;
+
+            // Gets the screen info dictionary (window, platform, display)
+            let dict = present_methods
+                .downcast_bound::<PyDict>(py)?
+                .get_item("screen")?
+                .ok_or(FragmentColorError::new_err("Object can't render to screen"))?;
+            let screen_info = dict.downcast::<PyDict>()?;
+
+            // Mandatory WindowHandle for all platforms
+            let window: u64 = screen_info
+                .get_item("window")?
+                .ok_or(FragmentColorError::new_err("Missing window handle"))?
+                .extract()?;
+            // Optional platform and display (only present on Linux)
+            let platform: String = screen_info
+                .get_item("platform")?
+                .unwrap_or("".into_pyobject(py)?.into_any())
+                .extract()?;
+            let display: u64 = screen_info
+                .get_item("display")?
+                .unwrap_or(0u64.into_pyobject(py)?.into_any())
+                .extract()?;
+
+            // Gets the window size to configure the surface
+            let size: (u32, u32) = rendercanvas
+                .call_method0(py, "get_physical_size")?
+                .downcast_bound(py)?
+                .extract()?;
+
+            let size = wgpu::Extent3d {
+                width: size.0,
+                height: size.1,
+                depth_or_array_layers: 1,
             };
 
-            Ok(dict.unbind())
+            let (window_handle, display_handle) =
+                create_raw_handles(platform, window, Some(display))?;
+
+            let handle = PyWindowHandle {
+                window_handle,
+                display_handle,
+            };
+
+            let (context, surface, config) = pollster::block_on(self.create_surface(handle, size))?;
+            target.init(context, surface, config);
+
+            Ok(target.into_pyobject(py)?.unbind())
         })
     }
-}
-
-impl RenderCanvasContext {
-    pub(crate) fn init_context(&mut self, renderer: Py<Renderer>, target: Py<RenderCanvasTarget>) {
-        self.renderer = Some(renderer);
-        self.target = Some(target);
-    }
-}
-
-#[pyfunction]
-/// The context hook that will be called from Python by RenderCanvas
-/// When the user calls `RenderCanvas.get_context("fragmentcolor")`
-pub fn rendercanvas_context_hook(
-    canvas: PyObject,
-    present_methods: PyObject,
-) -> RenderCanvasContext {
-    RenderCanvasContext::new(canvas, present_methods)
 }
 
 /// A Python module implemented in Rust. The name of this function must match
@@ -125,7 +152,6 @@ pub fn fragmentcolor(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(rendercanvas_context_hook, m)?)?;
     m.add_class::<RenderCanvasTarget>()?;
     m.add_class::<RenderCanvasFrame>()?;
-    m.add_class::<RenderCanvasContext>()?;
 
     // Custom error type
     m.add(
